Move CLI to standalone package, add tests, remove old scripts/cli
- Moved CLI from scripts/cli.ts to packages/cli/ standalone package
- Tests now live at packages/cli/src/index.test.ts (43 tests, all pass)
- Removed scripts/cli.ts, scripts/cli.test.ts, and `npm run cli` script
- Added `npm run cli:test` convenience script
