+++ package.json
@@ -42,10 +42,10 @@
     "mcp:dev:status": "node ./scripts/enode.js ./node_modules/.bin/tsx scripts/mcp-dev.ts status",
     "install:local": "node ./scripts/install-local.js",
     "sign:win": "powershell -NoProfile -ExecutionPolicy Bypass -File ./scripts/sign-windows.ps1",
-    "cli": "node ./scripts/enode.js ./node_modules/.bin/tsx scripts/cli.ts",
     "detect-patterns": "node ./scripts/enode.js ./node_modules/.bin/tsx scripts/detect-patterns.ts",
     "mcpb:pack": "npx @anthropic-ai/mcpb pack mcpb/",
     "cli:build": "cd packages/cli && npm run build",
+    "cli:test": "cd packages/cli && npm test",
     "cli:link": "cd packages/cli && npm link"
   },
   "keywords": [],
